+++ CHANGES.md
@@ -62,6 +62,9 @@
 * Change hover color of 'Logout', 'Restart' and 'Shutdown' menu items
 * Change ensure alert banners don't overlap submenus on view-show
 * Change remove trailing slash in void elements (img, br, hr, input, etc.) as not needed in HTML5
+* Change remaining submenus from partial JS to full Python to remove button display lag
+* Change move "About" page from /config/about to root /about
+* Fix hide the media process config button on the submenu of the media process config page 
 
 
 [develop]